rename mapStateInteraction to mapState
