+++ app/src/app/_components/layouts/Header/User/UserLoggedIn.tsx
@@ -6,7 +6,7 @@ import { Fragment } from 'react'
 import { Link } from 'src/app/_components/links/Link'
 import { useHasPermissions } from 'src/app/_hooks/useHasPermissions'
 import { useRegionSlug } from 'src/app/regionen/[regionSlug]/_components/regionUtils/useRegionSlug'
-import { useMapStoreActions } from 'src/app/regionen/[regionSlug]/_hooks/mapStateInteraction/useMapStateInteraction'
+import { useMapStoreActions } from 'src/app/regionen/[regionSlug]/_hooks/mapState/useMapState'
 import logout from 'src/auth/mutations/logout'
 import { isAdmin } from 'src/users/components/utils/usersUtils'
 import { CurrentUser } from 'src/users/queries/getCurrentUser'
