+++ app/src/app/regionen/[regionSlug]/_components/DebugBoxes/DebugMap.tsx
@@ -3,8 +3,8 @@ import { useState } from 'react'
 import { useMap } from 'react-map-gl/maplibre'
 import { getTilesUrl, isDevTilesUrl } from 'src/app/_components/utils/getTilesUrl'
 import { twJoin } from 'tailwind-merge'
-import { useMapDebugState } from '../../_hooks/mapStateInteraction/useMapDebugState'
-import { useMapStoreLoaded } from '../../_hooks/mapStateInteraction/useMapStateInteraction'
+import { useMapDebugState } from '../../_hooks/mapState/useMapDebugState'
+import { useMapStoreLoaded } from '../../_hooks/mapState/useMapState'
 import { useInteractiveLayers } from '../Map/utils/useInteractiveLayers'
 import { DebugMapDownload } from './DebugMapDownload'
