+++ app/src/app/regionen/[regionSlug]/_components/Map/SourcesAndLayers/LayerHighlight.tsx
@@ -4,7 +4,7 @@ import {
   useMapStoreLoaded,
   useMapStoreInspectorFeatures,
   useMapStoreCalculatorAreasWithFeatures,
-} from '../../../_hooks/mapStateInteraction/useMapStateInteraction'
+} from '../../../_hooks/mapState/useMapState'
 import {
   useSelectedFeatures,
   type SelectedFeature,
