+++ app/src/app/regionen/[regionSlug]/_components/Map/SourcesAndLayers/SourcesLayersAtlasGeo.tsx
@@ -5,7 +5,7 @@ import { makeTileUrlCacheless } from 'src/app/_components/utils/getTilesUrl'
 import {
   Store,
   useMapDebugState,
-} from 'src/app/regionen/[regionSlug]/_hooks/mapStateInteraction/useMapDebugState'
+} from 'src/app/regionen/[regionSlug]/_hooks/mapState/useMapDebugState'
 import { useBackgroundParam } from 'src/app/regionen/[regionSlug]/_hooks/useQueryState/useBackgroundParam'
 import { useCategoriesConfig } from 'src/app/regionen/[regionSlug]/_hooks/useQueryState/useCategoriesConfig/useCategoriesConfig'
 import { debugLayerStyles } from 'src/app/regionen/[regionSlug]/_mapData/mapDataSubcategories/mapboxStyles/debugLayerStyles'
