+++ app/src/app/regionen/[regionSlug]/_components/SidebarInspector/InspectorFeatureAtlasNote/EditNoteForm.tsx
@@ -11,7 +11,7 @@ import deleteNote from 'src/notes/mutations/deleteNote'
 import updateNote from 'src/notes/mutations/updateNote'
 import getNoteAndComments, { NoteAndComments } from 'src/notes/queries/getNoteAndComments'
 import { twJoin, twMerge } from 'tailwind-merge'
-import { useMapStoreActions } from '../../../_hooks/mapStateInteraction/useMapStateInteraction'
+import { useMapStoreActions } from '../../../_hooks/mapState/useMapState'
 import { useQueryKey } from '../../notes/AtlasNotes/utils/useQueryKey'
 import { useStaticRegion } from '../../regionUtils/useStaticRegion'
 import { SvgNotesCheckmark } from '../icons/SvgNotesCheckmark'
