+++ app/src/app/regionen/[regionSlug]/_components/SidebarInspector/InspectorFeatureOsmNote.tsx
@@ -8,7 +8,7 @@ import { getOsmUrl } from 'src/app/_components/utils/getOsmUrl'
 import { isDev } from 'src/app/_components/utils/isEnv'
 import { ObjectDump } from 'src/app/admin/_components/ObjectDump'
 import { twJoin } from 'tailwind-merge'
-import { useOsmNotesFeatures } from '../../_hooks/mapStateInteraction/userMapNotes'
+import { useOsmNotesFeatures } from '../../_hooks/mapState/userMapNotes'
 import { Disclosure } from './Disclosure/Disclosure'
 import { InspectorOsmNoteFeature } from './Inspector'
 import { OsmUserLink } from './OsmUserLink'
