+++ app/src/app/regionen/[regionSlug]/_components/SidebarInspector/Verification/VerificationForm/VerificationForm.tsx
@@ -4,7 +4,7 @@ import { Suspense } from 'react'
 import { useRegionSlug } from 'src/app/regionen/[regionSlug]/_components/regionUtils/useRegionSlug'
 import { Spinner } from 'src/app/_components/Spinner/Spinner'
 import Form, { FORM_ERROR } from 'src/app/_components/forms/Form'
-import { useMapStoreActions } from 'src/app/regionen/[regionSlug]/_hooks/mapStateInteraction/useMapStateInteraction'
+import { useMapStoreActions } from 'src/app/regionen/[regionSlug]/_hooks/mapState/useMapState'
 import createBikelaneVerification from 'src/bikelane-verifications/mutations/createBikelaneVerification'
 import {
   FormVerificationSchema,
