+++ app/src/app/regionen/[regionSlug]/_components/notes/NotesNew/NotesNewModal.tsx
@@ -1,7 +1,7 @@
 import { Dialog, Transition, TransitionChild } from '@headlessui/react'
 import { Fragment, useRef } from 'react'
 import { CloseButton } from 'src/app/_components/CloseButton/CloseButton'
-import { useOsmNotesActions } from '../../../_hooks/mapStateInteraction/userMapNotes'
+import { useOsmNotesActions } from '../../../_hooks/mapState/userMapNotes'
 import { useNewAtlasNoteMapParam } from '../../../_hooks/useQueryState/useNotesAtlasParams'
 import { useNewOsmNoteMapParam } from '../../../_hooks/useQueryState/useNotesOsmParams'
