+++ app/src/app/regionen/[regionSlug]/_components/notes/OsmNotes/OsmNotesNewForm.tsx
@@ -5,8 +5,8 @@ import { buttonStylesOnYellow } from 'src/app/_components/links/styles'
 import { appBaseUrl } from 'src/app/_components/utils/appBaseUrl.const'
 import { getOsmApiUrl } from 'src/app/_components/utils/getOsmUrl'
 import { useHasPermissions } from 'src/app/_hooks/useHasPermissions'
-import { useMapStoreBounds } from 'src/app/regionen/[regionSlug]/_hooks/mapStateInteraction/useMapStateInteraction'
-import { useOsmNewNoteFeature } from '../../../_hooks/mapStateInteraction/userMapNotes'
+import { useMapStoreBounds } from 'src/app/regionen/[regionSlug]/_hooks/mapState/useMapState'
+import { useOsmNewNoteFeature } from '../../../_hooks/mapState/userMapNotes'
 import { osmTypeIdString, osmUrl } from '../../SidebarInspector/Tools/osmUrls/osmUrls'
 import { useRegion } from '../../regionUtils/useRegion'
 import { OsmNotesThread } from './types'
