+++ app/src/app/regionen/[regionSlug]/_components/utils/sourceKeyUtils/createInspectorFeatureKey.ts
@@ -1,4 +1,4 @@
-import { StoreFeaturesInspector } from '../../../_hooks/mapStateInteraction/useMapStateInteraction'
+import { StoreFeaturesInspector } from '../../../_hooks/mapState/useMapState'
 
 export const createInspectorFeatureKey = (
   feature: StoreFeaturesInspector['inspectorFeatures'][number],
