+++ app/src/app/regionen/[regionSlug]/_hooks/mapState/useMapDebugState.ts
